Add owner and pagination

+++ src/services/contactsService.js
@@ -1,25 +1,51 @@
 const Contacts = require("../schemas/contactsSchema");
 
-const apiListContacts = () => {
-  return Contacts.find();
+const apiListContacts = async (owner, { page, limit, favorite }) => {
+  const skip = (page - 1) * limit;
+  const contacts = await Contacts.find({ owner })
+    .select({ owner: 0 })
+    .skip(skip)
+    .limit(parseInt(limit));
+  if (favorite) {
+    return contacts.filter((contact) => Boolean(contact.favorite));
+  }
+  return contacts;
 };
 
-const apiGetContactById = (contactId) => {
-  return Contacts.findOne({ _id: contactId });
+const apiGetContactById = async (contactId, owner) => {
+  const contact = await Contacts.findById(contactId);
+  if (contact.owner.toString() !== owner) {
+    return null;
+  }
+  return contact;
 };
 
-const apiAddContact = (body) => {
-  return Contacts.create(body);
+const apiAddContact = async (body, owner) => {
+  const newContact = await Contacts.create({ ...body, owner });
+  return newContact;
 };
 
-const apiUpdateContact = (contactId, body) => {
+const apiUpdateContact = async (contactId, body, owner) => {
+  const contact = await Contacts.findById(contactId);
+  if (contact.owner.toString() !== owner) {
+    return null;
+  }
   return Contacts.findByIdAndUpdate(contactId, body, { new: true });
 };
-const apiUpdateStatusContact = (contactId, body) => {
+
+const apiUpdateStatusContact = async (contactId, body, owner) => {
+  const contact = await Contacts.findById(contactId);
+  if (contact.owner.toString() !== owner) {
+    return null;
+  }
   return Contacts.findByIdAndUpdate(contactId, body);
 };
 
-const apiRemoveContact = (contactId) => {
+const apiRemoveContact = async (contactId, owner) => {
+  const contact = await Contacts.findById(contactId);
+  if (contact.owner.toString() !== owner) {
+    return null;
+  }
   return Contacts.findByIdAndRemove(contactId);
 };
